Moving make_fov into new binplotlib from individual plotting binaries
