+++ codebase/superdarn/src.bin/tk/plot/fov_plot.1.10/fov_plot.c
@@ -70,6 +70,7 @@
 
 #include "radar.h" 
 
+#include "griddata.h"
 #include "make_fov.h"
 #include "text_box.h"
 #include "plot_time_label.h"
