+++ codebase/superdarn/src.bin/tk/plot/grid_plot.1.12/grid_plot.c
@@ -921,7 +921,7 @@ int main(int argc,char *argv[]) {
   if ((lat>0) && (latmin<0)) latmin=-latmin;
 
   if ((fovflg || ffovflg) && !gfovflg) {
-    fov=make_fov(rgrid->st_time,network,chisham,old_aacgm);
+    fov=make_grid_fov(rgrid->st_time,network,chisham,old_aacgm);
     if (!magflg) {
       if (old_aacgm) MapModify(fov,AACGMtransform,&flg);
       else           MapModify(fov,AACGM_v2_transform,&flg);
@@ -1200,7 +1200,7 @@ int main(int argc,char *argv[]) {
     if (avflg) GridAverage(rgrid,rgridavg,aval+cprm*(aval !=0)); 
 
     if ((fovflg || ffovflg) && gfovflg) {
-      fov=make_fov_data(rgrid,network,chisham,old_aacgm);
+      fov=make_grid_fov_data(rgrid,network,chisham,old_aacgm);
       if (!magflg) {
         if (old_aacgm) MapModify(fov,AACGMtransform,&flg);
         else           MapModify(fov,AACGM_v2_transform,&flg);
