+++ codebase/superdarn/src.bin/tk/plot/grid_plot.1.12/makefile
@@ -5,10 +5,9 @@ include $(MAKECFG).$(SYSTEM)
 
 INCLUDE=-I$(IPATH)/base -I$(IPATH)/general -I$(IPATH)/analysis -I$(IPATH)/imagery -I$(IPATH)/superdarn
 
-SRC = grid_plot.c hlpstr.h plot_time.c plot_time.h \
-      errstr.h make_fov.c make_fov.h
+SRC = grid_plot.c hlpstr.h plot_time.c plot_time.h errstr.h 
 
-OBJS = plot_time.o make_fov.o grid_plot.o 
+OBJS = plot_time.o grid_plot.o 
 
 DSTPATH = $(BINPATH)
 OUTPUT = grid_plot
