+++ codebase/superdarn/src.lib/tk/binplotlib.1.0/src/make_fov.c
@@ -35,13 +35,134 @@
 #include "radar.h"
 #include "rpos.h"
 #include "griddata.h"
-#include "cnvmap.h"
 #include "polygon.h"
 
 
-/* This function returns FOVs for all operational radars */
 struct PolygonData *make_fov(double tval,struct RadarNetwork *network,
-                             int chisham,int old_aacgm) {
+                             float alt,int chisham) {
+
+    double rho,lat,lon;
+    int i,rn,bm;
+    float pnt[2];
+    int yr,mo,dy,hr,mt;
+    double sc;
+    int frang=180;
+    int rsep=45;
+    struct PolygonData *ptr=NULL;
+    struct RadarSite *site=NULL;
+    int maxrange=75;
+
+    TimeEpochToYMDHMS(tval,&yr,&mo,&dy,&hr,&mt,&sc);
+
+    ptr=PolygonMake(sizeof(float)*2,NULL);
+
+    for (i=0;i<network->rnum;i++) {
+
+        site=RadarYMDHMSGetSite(&(network->radar[i]),yr,mo,dy,hr,mt,(int) sc);
+        if (site==NULL) continue;
+        PolygonAddPolygon(ptr,i);
+
+        for (rn=0;rn<=maxrange;rn++) {
+            RPosGeo(0,0,rn,site,frang,rsep,
+                    site->recrise,alt,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+
+        for (bm=1;bm<=site->maxbeam;bm++) {
+            RPosGeo(0,bm,maxrange,site,frang,rsep,
+                    site->recrise,alt,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+
+        for (rn=maxrange-1;rn>=0;rn--) {
+            RPosGeo(0,site->maxbeam,rn,site,frang,rsep,
+                    site->recrise,alt,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+
+        for (bm=site->maxbeam-1;bm>0;bm--) {
+            RPosGeo(0,bm,0,site,frang,rsep,
+                    site->recrise,alt,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+    }
+    return ptr;
+}
+
+
+struct PolygonData *make_field_fov(double tval,struct RadarNetwork *network,
+                                   int id,int chisham) {
+
+    double rho,lat,lon;
+    int i,rn,bm;
+    float pnt[2];
+    int yr,mo,dy,hr,mt;
+    double sc;
+    int frang=180;
+    int rsep=45;
+    struct PolygonData *ptr=NULL;
+    struct RadarSite *site=NULL;
+    int maxrange=75;
+
+    TimeEpochToYMDHMS(tval,&yr,&mo,&dy,&hr,&mt,&sc);
+
+    ptr=PolygonMake(sizeof(float)*2,NULL);
+
+    for (i=0;i<network->rnum;i++) {
+
+        if (network->radar[i].status !=1) continue;
+        if (network->radar[i].id==id) continue;
+        site=RadarYMDHMSGetSite(&(network->radar[i]),yr,mo,dy,hr,mt,(int) sc);
+        if (site==NULL) continue;
+        PolygonAddPolygon(ptr,1);
+
+        for (rn=0;rn<=maxrange;rn++) {
+            RPosGeo(0,0,rn,site,frang,rsep,
+                    site->recrise,0,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+
+        for (bm=1;bm<=site->maxbeam;bm++) {
+            RPosGeo(0,bm,maxrange,site,frang,rsep,
+                    site->recrise,0,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+
+        for (rn=maxrange-1;rn>=0;rn--) {
+            RPosGeo(0,site->maxbeam,rn,site,frang,rsep,
+                    site->recrise,0,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+
+        for (bm=site->maxbeam-1;bm>0;bm--) {
+            RPosGeo(0,bm,0,site,frang,rsep,
+                    site->recrise,0,&rho,&lat,&lon,chisham);
+            pnt[0]=lat;
+            pnt[1]=lon;
+            PolygonAdd(ptr,pnt);
+        }
+    }
+    return ptr;
+}
+
+
+/* This function returns FOVs for all operational radars */
+struct PolygonData *make_grid_fov(double tval,struct RadarNetwork *network,
+                                  int chisham,int old_aacgm) {
 
     double rho,lat,lon;
     int i,rn,bm;
@@ -106,9 +227,9 @@ struct PolygonData *make_fov(double tval,struct RadarNetwork *network,
 }
 
 
-/* This function returns only the radar FoVs that are contributing grid vectors */
-struct PolygonData *make_fov_data(struct GridData *gptr,struct RadarNetwork *network,
-                                  int chisham,int old_aacgm) {
+/* This function returns only the radar FOVs that are contributing grid vectors */
+struct PolygonData *make_grid_fov_data(struct GridData *gptr,struct RadarNetwork *network,
+                                       int chisham,int old_aacgm) {
 
     double rho,lat,lon;
     int i,j,rn,bm;
